requirements.txt
pandas
streamlit
scikit-learn
tensorflow-cpu